+++ packages/react/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-07-21
+
+### Added
+
+- Controlled `open`/`searchQuery`, change callbacks, loading/invalid callbacks, and reactive runtime options.
+
 ## [0.3.0] - 2026-07-21
 
 ### Added
@@ -34,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release: `ForgeSelectReact` component wrapping the `forge-select` core library, with controlled `value`/`onChange` support.
 
-[Unreleased]: https://github.com/cmm-cmm/ForgeSelect/compare/react-v0.3.0...HEAD
+[Unreleased]: https://github.com/cmm-cmm/ForgeSelect/compare/react-v0.4.0...HEAD
+[0.4.0]: https://github.com/cmm-cmm/ForgeSelect/compare/react-v0.3.0...react-v0.4.0
 [0.3.0]: https://github.com/cmm-cmm/ForgeSelect/compare/react-v0.2.0...react-v0.3.0
 [0.2.0]: https://github.com/cmm-cmm/ForgeSelect/compare/react-v0.1.0...react-v0.2.0
 [0.1.0]: https://github.com/cmm-cmm/ForgeSelect/releases/tag/react-v0.1.0